Remove SwiftSyntaxParser module
This removes the `SwiftSyntaxParser` module, which was just a wrapper around `SwiftParser` to ease the transition for clients that previously used `SwiftSyntaxParser`. Everybody should be using `SwiftParser` by now, so remove the deprecated module.

This migrates all test cases that still provide value to other modules.
