--- Sources/SwiftSyntaxParser/Documentation.docc/Index.md
@@ -1,3 +1,0 @@
-# ``SwiftSyntaxParser``
-
-SwiftSyntaxParser allows parsing Swift source code into a SwiftSyntax tree.
