--- Tests/SwiftSyntaxParserTest/Inputs/closure.swift
@@ -1,3 +1,0 @@
-// A closure without a signature. The test will ensure it stays the same after
-// applying a rewriting pass.
-let x: () -> Void = {}
