--- Tests/SwiftSyntaxParserTest/Inputs/diagnostics.swift
@@ -1,11 +1,0 @@
-func foo() {
-
-}
-
-func bar() {
-
-}
-
-func baz() {
-
-}
