--- Tests/SwiftSyntaxParserTest/Inputs/near-empty.swift
@@ -1,1 +1,0 @@
-n
