--- Tests/SwiftSyntaxParserTest/Inputs/nested-blocks.swift
@@ -1,8 +1,0 @@
-struct Foo {
-  func foo() {
-    print("hello")
-    func bar() {
-      print("goodbye")
-    }
-  }
-}
