--- Tests/SwiftSyntaxParserTest/Inputs/visitor.swift
@@ -1,7 +1,0 @@
-func foo() {
-  public func foo() {
-    func foo() {
-      /*Unknown token */0xG
-    }
-  }
-}
